% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fintxt_client_get}
\alias{fintxt_client_get}
\title{Retrieve historic or live news intensities}
\usage{
fintxt_client_get(endpoint, language, ric, date = NULL)
}
\arguments{
\item{endpoint}{API endpoint to use. Must be one of must be one of 'languages', 'live', 'historic'}

\item{language}{Filter news intensity values by language. See the '/languages' endpoint for allowed values}

\item{ric}{RIC code for the company for which you want to query news intensity values}

\item{date}{Filter news intensity values by date. Defaults to NULL.}
}
\description{
Retrieve historic or live news intensities
}
