% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fintxt_get_languages}
\alias{fintxt_get_languages}
\title{Retrieve languages for which you can query news intensities}
\usage{
fintxt_get_languages()
}
\value{
list of languages
}
\description{
Retrieve languages for which you can query news intensities
}
\examples{
\dontrun{
# Call languages endpoint
library(FinTxtClient)
lang <- fintxt_get_languages()
}

}
