% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fintxt_historic_intensities_portfolio}
\alias{fintxt_historic_intensities_portfolio}
\title{Retrieve historic news intensities for a portfolio of company RICs or commodities}
\usage{
fintxt_historic_intensities_portfolio(type = c("companies", "commodities"),
  language = c("total", "english", "french", "russian", "arabic", "german"),
  date, identifiers, weights)
}
\arguments{
\item{type}{either one of 'companies' or 'commodities'}

\item{language}{Filter news intensity values by language. See the '/languages' endpoint for allowed values.}

\item{date}{Filter news intensity values by date.}

\item{identifiers}{RIC codes for the companies for which you want to query news intensity values or names of commodities for which you want to query news intensity values. See documentation for supported commodities. Note that you cannot mix companies and commodities}

\item{weights}{Weight of each company/commodity in your portfolio in decimal format. Should sum to 1, but the API will still return news intensities if this is not the case. See example section.}
}
\value{
List containing API results
}
\description{
Retrieve historic news intensities for a language, a date and a company's \href{https://en.wikipedia.org/wiki/Reuters_Instrument_Code}{Reuters Instrument Code} or commodities. Note that you need to supply a valid api token.
}
\examples{
\dontrun{
# historic portfolio endpoint
identifiers = c("TRI.TO", "IBM.N", "RRD.N", "SPGI.N", "INTU.OQ", "RELN.AS", "WLSNc.AS", "REL.L"),
weights = c(0.3, 0.1,0.05,0.05, 0.2,0.1,0.1,0.1)
fintxt_historic_intensities_portfolio("companies", "english", "18-06-2018", identifiers, weights)
}

}
\seealso{
See the documentation at <https://fintxt.github.io/documentation/index.html>
}
