% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fintxt_live_intensities}
\alias{fintxt_live_intensities}
\title{Retrieve live news intensities}
\usage{
fintxt_live_intensities(language, ric)
}
\arguments{
\item{language}{Filter news intensity values by language. See the '/languages' endpoint for allowed values.}

\item{ric}{RIC code for the company for which you want to query news intensity values.}
}
\value{
List containing API results
}
\description{
Retrieve live news intensities for a language and a company's \href{https://en.wikipedia.org/wiki/Reuters_Instrument_Code}{Reuters Instrument Code}. Note that you need to supply a valid api token.
}
\examples{
\dontrun{
# Call languages endpoint
library(FinTxtClient)
r <- fintxt_live_intensities("english", "TRI.TO")
}

}
\seealso{
See the documentation at <https://fintxt.github.io/documentation/index.html>
}
