% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fintxt_live_intensities_one}
\alias{fintxt_live_intensities_one}
\title{Retrieve live news intensities}
\usage{
fintxt_live_intensities_one(type = c("companies", "commodities"),
  language = c("total", "english", "french", "russian", "arabic", "german"),
  q)
}
\arguments{
\item{type}{either one of 'companies' or 'commodities'}

\item{language}{Filter news intensity values by language. See the '/languages' endpoint for allowed values.}

\item{q}{RIC code for the company for which you want to query news intensity values or name of commodity for which you want to query news intensity values. See documentation for supported commodities.}
}
\value{
List containing API results
}
\description{
Retrieve live news intensities for a language and a company's \href{https://en.wikipedia.org/wiki/Reuters_Instrument_Code}{Reuters Instrument Code}. Note that you need to supply a valid api token.
}
\examples{
\dontrun{
# Call languages endpoint
library(FinTxtClient)
r <- fintxt_live_intensities("english", "TRI.TO")
}

}
\seealso{
See the documentation at <https://fintxt.github.io/documentation/index.html>
}
